Moved calculators to  separate submodule
